requirements.txt
slack_bolt
requests
python-taiga